feat: add transaction listing component
- Creates a reusable component to list transactions
