+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "budget-management",
   "private": true,
-  "version": "1.7.0",
+  "version": "1.8.0",
   "type": "module",
   "scripts": {
     "dev": "vite",
